Move the "localhost = multiple instances" hint into the adhoc server selector
