+++ UI/MiscViews.h
@@ -90,4 +90,10 @@ protected:
 	uint32_t color_ = 0xFFC0C0C0;
 };
 
+class SettingHint : public UI::TextView {
+public:
+	SettingHint(std::string_view text);
+	void Draw(UIContext &dc) override;
+};
+
 void AddRotationPicker(ScreenManager *screenManager, UI::ViewGroup *parent, bool text);
